Merge with the Experimental PList branch.
Includes the new PropertyListKFS class, along with a global module and a subclass for reading and and writing Apple Property List files.

Although all of the included items in this merge are not considered experimental anymore, there is still much more work to do.  Lots of reinforcing, commenting, and documenting needs to be done.

Some other classes included due to dependencies:
  OrderedDictionaryKFS
  SimpleBigStringKFS
  SwapGlobalsKFS
  Logging Framework
  Select functions from General Global Modules

PropertyListKFS is a subclass of OrderedDictionaryKFS.
PropertyListKFS uses SimpleBigStringKFS for file IO.
SimpleBigStringKFS uses SwapGlobalsKFS for managing swap files.
SwapGlobalsKFS uses the Logging Framework for reporting things.
The above classes use a handful of functions from the General Global Modules.