Very weird runtime error when compiled with RB 2010r2.1.
RB 2010r2.1; Problem only visible in debug mode.
An exception is raised, but the encolsing try block doesn't catch it.  Instead, it goes all the way to App.UnhandledException.

The exception is raised inside TestBigStringKFS.TestModifyValue.  Modifying the length of the code of any method in that class after and not including the TestModifyValue method causes the problem to mysteriously go away (after meaning below, in the listbox of methods in the IDE).

Checked a Core 2 Duo MBP running 10.6.4 (yes bug), and a Dual 1 GHz Power Mac G4 running 10.5.8 (no bug).