Fixed the swap file leak in SwapGlobalsKFS.
ReleaseSwapFile was removing items from the pool even when deleting the file failed.

Because the most common error was File In Use, keeping the file in the pool when it can't be deleted immediately causes the App class to try and delete the file again when the app quits.  The stream that was accessing the file is most likely long gone by the time the App's Close event fires.