Found a bug in the compiler for RB 2010r3.
An exception is raised, and is not caught by the enclosing Try block.  Adding or removing a line of code in any location after the first Dim statement in the first constructor in the BigStringKFS class causes the problem to go away.

Tested on a Core 2 Duo MBP running OS X 10.6.4.