ProgressDelegateKFS: Rewrote the children data structure.
Now uses add and remove methods, rather than add and garbage collect methods.

This gets around a bug with WeakRefs in RB where they are not trustable while the target's Destructor is running.

On a side note, this implementation is actually faster and easier to maintain than the previous version.