UnitTesting: Use strong delegates for the post processing hooks.
Not sure if this is a good idea yet.

This has a side effect of preventing an Arbiter from deallocating until all the working threads have returned.

There is currently no way to detect if a weak delegate has gone dead, so when it happens, you get a NilObjectException.  With this patch, the error does not happen.  Not sure about that side effect, though.